+++ PayPalWebPayments/src/main/java/com/paypal/android/paypalwebpayments/PayPalWebVaultRequest.kt
@@ -1,0 +1,12 @@
+package com.paypal.android.paypalwebpayments
+
+/**
+ * Request to vault a PayPal payment method using [PayPalWebCheckoutClient.vault].
+ *
+ * @property [setupTokenId] ID for the setup token associated with the vault approval
+ * @property [approveVaultHref] URL for the approval web page
+ */
+data class PayPalWebVaultRequest(
+    val setupTokenId: String,
+    val approveVaultHref: String
+)